Made it cleaner, but it doesn't work now
Added a .env file, which contains the firebase's secret stuff and which is also in the .gitignore. Also instead of document.title = 'smth' I used svelte:head, because idk. Well the auth doesn't work now: in the /signup when I try to create an account, it says 'auth is not defined', but when I define auth by basicly rewriting the whole Firebase.js (which is now btw isn't in the .gitignore), it says some stupid shit like 'two fuckers have the same outpot at node_modules/some-shit'. idk